Update CHANGELOG
Corrections to MC-PDFT availabilities and method names (they should all have hyphens). MS-PDFT is not a method by itself; rather, it is an abstraction of the steps that lie in common between CMS-PDFT and XMS-PDFT (and several other methods not currently implemented). Dipoles are currently only available for MC-PDFT and CMS-PDFT, and transition dipoles only for the latter.